✅removed the download of pre-trained models from local tests
