+++ tests/compilation/test_predictor_rl.py
@@ -11,17 +11,8 @@ from mqt.predictor import reward, rl
 IN_GITHUB_ACTIONS = os.getenv("GITHUB_ACTIONS") == "true"
 
 
-@pytest.mark.parametrize(
-    "figure_of_merit",
-    ["expected_fidelity", "critical_depth"],
-)
-def test_qcompile_with_pretrained_models(figure_of_merit: reward.figure_of_merit) -> None:
-    qc = get_benchmark("ghz", 1, 3)
-    qc_compiled, compilation_information = rl.qcompile(qc, figure_of_merit=figure_of_merit, device_name="ionq_harmony")
-    assert isinstance(qc_compiled, QuantumCircuit)
-    assert compilation_information is not None
-
-
+# only run test when executed on GitHub runner
+@pytest.mark.skipif(not IN_GITHUB_ACTIONS, reason="Only run this test when executed on GitHub runner")
 @pytest.mark.parametrize(
     "figure_of_merit",
     ["expected_fidelity", "critical_depth"],
@@ -45,6 +36,6 @@ def test_qcompile_with_newly_trained_models(figure_of_merit: reward.figure_of_me
 def test_qcompile_with_false_input() -> None:
     qc = get_benchmark("dj", 1, 5)
     with pytest.raises(ValueError, match="figure_of_merit must not be None if predictor_singleton is None."):
-        rl.helper.qcompile(qc, None, "ibm_washington")
+        rl.helper.qcompile(qc, None, "quantinuum_h2")
     with pytest.raises(ValueError, match="device_name must not be None if predictor_singleton is None."):
         rl.helper.qcompile(qc, "expected_fidelity", None)
